Split `Namespace` into a class and an interface
